[WIP] test handleCustomerChange
