+++ src/__tests__/TimeInputForm.spec.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { shallow, mount } from "enzyme";
+import TimeInputForm from "../components/TimeInputForm";
+
+describe("<TimeInputForm />", () => {
+	it("calls timeInputForm from SaveTimeRecording", () => {
+		TimeInputForm.timeInputForm = jest.fn();
+		const describedComponent = mount(<TimeInputForm />);
+		describedComponent.find({rate: "50"});
+		describedComponent.instance();
+		expect(describedComponent.find("Input"))
+	})
+
+	it("has 3 input fields", () => {
+		const describedComponent = shallow(<TimeInputForm />);
+		expect(describedComponent.find("Input")).toHaveLength(3);
+	})
+})
